# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Jakarta EV Charging Stations API",
    "description": "\nCharging-station data for Jakarta and the rest of Indonesia. It combines the official PLN\nSPKLU registry (petaspklu.id) with Open Charge Map and OpenStreetMap.\n\nUse it to list and filter stations, find what's nearby, get GeoJSON for the map, plan a\nroute to a charger, and read summary stats.\n",
    "contact": {
      "name": "EV-FLOW",
      "email": "softopen24@gmail.com"
    },
    "license": {
      "name": "Data: PLN, OCM (CC-BY-SA), OSM (ODbL)"
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/stations": {
      "get": {
        "tags": [
          "stations"
        ],
        "summary": "List / filter charging stations",
        "operationId": "list_stations_api_v1_stations_get",
        "parameters": [
          {
            "name": "source",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Source"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by dataset.",
              "title": "Source"
            },
            "description": "Filter by dataset."
          },
          {
            "name": "province",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Exact province match (case-insensitive), e.g. 'DKI Jakarta'.",
              "title": "Province"
            },
            "description": "Exact province match (case-insensitive), e.g. 'DKI Jakarta'."
          },
          {
            "name": "city",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "City/kabupaten substring match.",
              "title": "City"
            },
            "description": "City/kabupaten substring match."
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Case-insensitive search on station name.",
              "title": "Q"
            },
            "description": "Case-insensitive search on station name."
          },
          {
            "name": "min_power",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number",
                  "minimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "description": "Min power (kW).",
              "title": "Min Power"
            },
            "description": "Min power (kW)."
          },
          {
            "name": "max_power",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number",
                  "minimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "description": "Max power (kW).",
              "title": "Max Power"
            },
            "description": "Max power (kW)."
          },
          {
            "name": "connector_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Connector standard, e.g. 'CCS2' or 'AC Type 2' (see /api/v1/connectors). Currently inferred.",
              "examples": [
                "CCS2"
              ],
              "title": "Connector Type"
            },
            "description": "Connector standard, e.g. 'CCS2' or 'AC Type 2' (see /api/v1/connectors). Currently inferred."
          },
          {
            "name": "speed_tier",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Speed tier: slow / medium / fast / ultra_fast (see /api/v1/speed-tiers).",
              "title": "Speed Tier"
            },
            "description": "Speed tier: slow / medium / fast / ultra_fast (see /api/v1/speed-tiers)."
          },
          {
            "name": "bbox",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Bounding box 'minLon,minLat,maxLon,maxLat'.",
              "examples": [
                "106.55,-6.65,107.10,-5.95"
              ],
              "title": "Bbox"
            },
            "description": "Bounding box 'minLon,minLat,maxLon,maxLat'."
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "description": "Page size.",
              "default": 100,
              "title": "Limit"
            },
            "description": "Page size."
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Page offset.",
              "default": 0,
              "title": "Offset"
            },
            "description": "Page offset."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StationList"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/stations/nearby": {
      "get": {
        "tags": [
          "stations"
        ],
        "summary": "Nearest stations to a point ('near me')",
        "operationId": "nearby_api_v1_stations_nearby_get",
        "parameters": [
          {
            "name": "lat",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "maximum": 90,
              "minimum": -90,
              "description": "Origin latitude.",
              "examples": [
                -6.2088
              ],
              "title": "Lat"
            },
            "description": "Origin latitude."
          },
          {
            "name": "lon",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "maximum": 180,
              "minimum": -180,
              "description": "Origin longitude.",
              "examples": [
                106.8456
              ],
              "title": "Lon"
            },
            "description": "Origin longitude."
          },
          {
            "name": "radius_km",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "maximum": 500,
              "exclusiveMinimum": 0,
              "description": "Search radius (km).",
              "default": 5.0,
              "title": "Radius Km"
            },
            "description": "Search radius (km)."
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "description": "Max results, sorted by distance.",
              "default": 20,
              "title": "Limit"
            },
            "description": "Max results, sorted by distance."
          },
          {
            "name": "source",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Source"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Optional source filter.",
              "title": "Source"
            },
            "description": "Optional source filter."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Station"
                  },
                  "title": "Response Nearby Api V1 Stations Nearby Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/stations/{station_id}": {
      "get": {
        "tags": [
          "stations"
        ],
        "summary": "Fetch one station by id",
        "operationId": "get_station_api_v1_stations__station_id__get",
        "parameters": [
          {
            "name": "station_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Station Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Station"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/ev-models": {
      "get": {
        "tags": [
          "ev-models"
        ],
        "summary": "List EV models (catalogue from the Kaggle Indonesia-EV-2026 dataset)",
        "operationId": "ev_models_api_v1_ev_models_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Case-insensitive search on vehicle name.",
              "title": "Q"
            },
            "description": "Case-insensitive search on vehicle name."
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "description": "Page size.",
              "default": 100,
              "title": "Limit"
            },
            "description": "Page size."
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Page offset.",
              "default": 0,
              "title": "Offset"
            },
            "description": "Page offset."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EVModelList"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/ev-models/{model_id}": {
      "get": {
        "tags": [
          "ev-models"
        ],
        "summary": "Fetch one EV model by id",
        "operationId": "ev_model_api_v1_ev_models__model_id__get",
        "parameters": [
          {
            "name": "model_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Model Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EVModel"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Source": {
        "type": "string",
        "enum": [
          "pln_spklu",
          "open_charge_map",
          "osm"
        ],
        "title": "Source"
      },
...
      "StationList": {
        "properties": {
          "total": {
            "type": "integer",
            "title": "Total",
            "description": "Total matching records (before pagination).",
            "examples": [
              1142
            ]
          },
          "limit": {
            "type": "integer",
            "title": "Limit",
            "examples": [
              100
            ]
          },
          "offset": {
            "type": "integer",
            "title": "Offset",
            "examples": [
              0
            ]
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/Station"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "total",
          "limit",
          "offset",
          "items"
        ],
        "title": "StationList"
      },
...
    }
  },
...
}